requirements.txt
Flask==3.0.3
google-generativeai==0.7.2  # Replace with the actual version if different
requests==2.32.3
httpx==0.27.0
pycryptodomex==3.20.0
mutagen==1.47.0
pytube==15.0.0
h11==0.14.0
httpcore==1.0.5
websockets==12.0
Werkzeug==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
